+++ NEWS.md
@@ -1,3 +1,15 @@
+Version 0.2.0
+=============
+
+Added function simdata, which generates simulated rating scale data and
+allows msd to be tested for accuracy.
+
+Added the option to specify a wider range of possible rating categories
+for functions pms and misfit than may be present in the data. This is
+important because pms and misfit are useful for estimating person
+measures and misfit statistics for individual persons, and it is
+possible that the person did not use all rating categories.
+
 Version 0.1.0
 =============